feat: cross-language fallback, tappable synonyms, and a data_version that distinguishes builds
Three buildable items plus four answers to open decisions, and one of them turned
into a design that stays on the roadmap.

The cross-language fallback fires only when the active language returned nothing
resembling what you typed: no exact match and nothing in the top coverage band.
Measured before choosing the threshold, on the two real packs: 0 of 400 common
Spanish headwords trigger it, so the normal case pays nothing, and 321 of 400
common English ones do. The 79 that do not -- break, man, go, line, bear -- are
genuinely in the Spanish pack, so not falling back is the right answer there.

It does NOT simply go "after", which is how I had designed it. If the active
language returned ten phonetic near-misses, putting the correct answer below all
of them leaves it off screen, which is the same as not having searched. Both sets
merge through the same comparator and the active language breaks ties AFTER
quality, not before.

Synonyms, antonyms and related words now carry their category as a heading and
their terms as links, same rule as D-159: abbreviated where the headword needs
the width, spelled out in the card where nothing competes. Only terms that exist
in the pack get painted -- the colour is the promise that it leads somewhere --
and a term the pack lacks is still shown, just plain. Two batched queries rather
than one, because the 64-key cap is a hard SQLite limit and glosses already
average 58.8 keys in English; merging them would have silently truncated exactly
the terms that are worth linking.

data_version was two things wearing one name: which dump the content came from,
and which build this is. That is why no single format worked. The dump date moves
to source_date and data_version becomes AAAAMMDDHHMM derived by the builder --
orders like a number, reads as a date without a converter, and distinguishes two
builds of the same dump. It is a reserved key now, because anything writable by
hand eventually goes unbumped, which is the bug this closes. It does not fit in a
32-bit Int, so PackMetadata.dataVersion widened to Long.

And the core pack stays a design, by request. Writing it found a trap that would
have sunk the implementation quietly: 93% of Spanish inflected forms belong to
verbs, form is 32.9 MB of a 71.7 MB pack, and a core selected by rank takes 91%
of that table. The 7.5 MB estimate only holds with frequency-based selection.
Whoever builds it by filtering on rank -- the obvious choice, since it is the
column already there -- would get tens of MB and conclude the idea was wrong.

D-168, D-169, D-170.

+++ dict-data/src/main/kotlin/cl/fadiaz/dictionary/data/PackFile.kt
@@ -214,7 +214,7 @@ class PackFile private constructor(
             langTarget = meta["lang_dst"],
             fuzzyProfile = FuzzyProfile.fromId(meta["fuzzy_profile"]),
             entryCount = meta.getValue("entry_count").toInt(),
-            dataVersion = meta.getValue("data_version").toInt(),
+            dataVersion = meta.getValue("data_version").toLong(),
             license = meta.getValue("license"),
             attribution = meta.getValue("attribution"),
             // `meta[...]` y no `getValue`: un pack anterior a D-138 no la trae y tiene que
